Full steps integration

+++ accommodation-form/src/components/Accommodation/MultipleImageUploader.tsx
@@ -1,8 +1,8 @@
-import { useState } from "react";
-
 interface MultipleImageUploaderProps {
-  handleUpdloadImages: (file: [File]) => void;
-  handleDeleteImage: (index: number) => void;
+  handleUpdloadImages?: (
+    file: Array<{ file: File; preview: string }>
+  ) => void | undefined;
+  handleDeleteImage?: (index: number) => void | undefined;
   images: Array<{ file: File; preview: string }>;
 }
 
@@ -11,9 +11,6 @@ const MultipleImageUploader: React.FC<MultipleImageUploaderProps> = ({
   handleDeleteImage,
   images = [],
 }) => {
-  // const [images, setImages] = useState<{ file: File; preview: string }[]>([]);
-
-  // Manejar la carga de archivos
   const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     if (e.target.files) {
       const filesArray = Array.from(e.target.files); // Convertir a array
@@ -30,38 +27,43 @@ const MultipleImageUploader: React.FC<MultipleImageUploaderProps> = ({
       });
 
       // Esperar a que todas las imágenes se procesen
-      Promise.all(newImages).then((images) => {
-        console.log(images);
-        handleUpdloadImages(images);
-      });
+      Promise.all(newImages).then(
+        (images: Array<{ file: File; preview: string }>) => {
+          console.log(images);
+          if (handleUpdloadImages) handleUpdloadImages(images);
+        }
+      );
     }
   };
 
-  // Eliminar una imagen de la previsualización
   const handleRemoveImage = (index: number) => {
-    handleDeleteImage(index);
+    if (handleDeleteImage) handleDeleteImage(index);
   };
 
   return (
     <>
-      <label
-        htmlFor="countries"
-        className="block mb-2  text-left text-sm font-medium text-white dark:text-white"
-      >
-        Subir Imágenes
-      </label>
+      {handleUpdloadImages && (
+        <label
+          htmlFor="countries"
+          className="block mb-2  text-left text-sm font-medium text-white dark:text-white"
+        >
+          Subir Imágenes
+        </label>
+      )}
 
       <div className="p-4 border rounded shadow">
         <label className="block text-gray-700 font-bold mb-2"></label>
-        <input
-          type="file"
-          multiple
-          accept="image/*"
-          onChange={handleFileChange}
-          className="mb-2"
-        />
 
-        {/* Mostrar vista previa */}
+        {handleUpdloadImages && (
+          <input
+            type="file"
+            multiple
+            accept="image/*"
+            onChange={handleFileChange}
+            className="mb-2"
+          />
+        )}
+
         <div className="flex gap-2 mt-3 flex-wrap">
           {images?.map(
             (img: { file: File; preview: string }, index: number) => (
@@ -71,12 +73,15 @@ const MultipleImageUploader: React.FC<MultipleImageUploaderProps> = ({
                   alt={`preview-${index}`}
                   className="w-32 h-32 object-cover border rounded"
                 />
-                <button
-                  className="absolute top-0 right-0 bg-red-500 text-white text-xs px-2 py-1 rounded-full"
-                  onClick={() => handleRemoveImage(index)}
-                >
-                  ✕
-                </button>
+
+                {handleDeleteImage && (
+                  <button
+                    className="absolute top-0 right-0 bg-red-500 text-white text-xs px-2 py-1 rounded-full"
+                    onClick={() => handleRemoveImage(index)}
+                  >
+                    ✕
+                  </button>
+                )}
               </div>
             )
           )}
